♻️(frontend) replace grommet TextArea with cunningham TextArea
In order to do the transition to the cunningham design system,
we replace the grommet TextArea component with the
cunningham TextArea components.
